Update from main repo.

+++ mobilecoin/cli.py
@@ -1,4 +1,5 @@
 import argparse
+from collections import defaultdict
 from decimal import Decimal
 import json
 import os
@@ -6,16 +7,17 @@ from pathlib import Path
 import subprocess
 
 from mnemonic import Mnemonic
+import segno
 
-from mobilecoin.client import (
-    Client,
+from .utility import (
     pmob2mob,
+    TRANSACTION_FEE,
+    try_int,
 )
+from .client import Client
 
-NETWORK = 'testnet'
-assert NETWORK in ['testnet', 'mainnet']
-MC_DATA = Path.home() / '.mobilecoin' / NETWORK
-LOG_LOCATION = MC_DATA / 'wallet_server_log.txt'
+
+config = json.loads(os.environ['MOBILECOIN_CONFIG'])
 
 
 class CommandLineInterface:
@@ -59,6 +61,10 @@ class CommandLineInterface:
         self.create_args.add_argument('-b', '--block', type=int,
                                       help='Block index at which to start the account. No transactions before this block will be loaded.')
 
+        self.rename_args = subparsers.add_parser('rename', help='Change account name.')
+        self.rename_args.add_argument('account_id', help='Account ID code.')
+        self.rename_args.add_argument('name', help='New account name.')
+
         self.import_args = subparsers.add_parser('import', help='Import an account.')
         self.import_args.add_argument('seed', help='Account seed phrase, seed file, or root entropy hex.')
         self.import_args.add_argument('-n', '--name', help='Account name.')
@@ -68,8 +74,11 @@ class CommandLineInterface:
         self.export_args = subparsers.add_parser('export', help='Export seed phrase.')
         self.export_args.add_argument('account_id', help='Account ID code.')
 
-        self.delete_args = subparsers.add_parser('delete', help='Delete an account from local storage.')
-        self.delete_args.add_argument('account_id', help='Account ID code.')
+        self.qr_args = subparsers.add_parser('qr', help='Show account address as a QR code')
+        self.qr_args.add_argument('account_id', help='Account ID code.')
+
+        self.remove_args = subparsers.add_parser('remove', help='Remove an account from local storage.')
+        self.remove_args.add_argument('account_id', help='Account ID code.')
 
         self.list_args = subparsers.add_parser('list', help='List accounts.')
 
@@ -77,8 +86,9 @@ class CommandLineInterface:
         self.history_args.add_argument('account_id', help='Account ID code.')
 
         self.send_args = subparsers.add_parser('send', help='Send a transaction.')
-        self.send_args.add_argument('from_account_id', help='Account ID to send from.')
-        self.send_args.add_argument('amount', help='Amount of MOB to send.', type=float)
+        self.send_args.add_argument('--build-only', action='store_true', help='Just build the transaction, do not submit it.')
+        self.send_args.add_argument('account_id', help='Account ID to send from.')
+        self.send_args.add_argument('amount', help='Amount of MOB to send.')
         self.send_args.add_argument('to_address', help='Address to send to.')
 
     def _load_account_prefix(self, prefix):
@@ -98,79 +108,73 @@ class CommandLineInterface:
             exit(1)
 
     def start(self, offline=False, bg=False):
-        if NETWORK == 'testnet':
-            wallet_server_command = ['./full-service-testnet']
-        elif NETWORK == 'mainnet':
-            wallet_server_command = ['./full-service-mainnet']
-
-        wallet_server_command += [
-            '--wallet-db', str(MC_DATA / 'wallet-db/wallet.db'),
-            '--ledger-db', str(MC_DATA / 'ledger-db'),
+        wallet_server_command = [
+            config['executable'],
+            '--ledger-db', config['ledger-db'],
+            '--wallet-db', config['wallet-db'],
         ]
         if offline:
-            wallet_server_command += [
-                '--offline',
-            ]
+            wallet_server_command += ['--offline']
         else:
-            if NETWORK == 'testnet':
-                wallet_server_command += [
-                    '--peer mc://node1.test.mobilecoin.com/',
-                    '--peer mc://node2.test.mobilecoin.com/',
-                    '--tx-source-url https://s3-us-west-1.amazonaws.com/mobilecoin.chain/node1.test.mobilecoin.com/',
-                    '--tx-source-url https://s3-us-west-1.amazonaws.com/mobilecoin.chain/node2.test.mobilecoin.com/',
-                ]
-            elif NETWORK == 'mainnet':
-                wallet_server_command += [
-                    '--peer', 'mc://node1.prod.mobilecoinww.com/',
-                    '--peer', 'mc://node2.prod.mobilecoinww.com/',
-                    '--tx-source-url', 'https://ledger.mobilecoinww.com/node1.prod.mobilecoinww.com/',
-                    '--tx-source-url', 'https://ledger.mobilecoinww.com/node2.prod.mobilecoinww.com/',
-                ]
+            for peer in config['peer']:
+                wallet_server_command += ['--peer', peer]
+            for tx_source_url in config['tx-source-url']:
+                wallet_server_command += ['--tx-source-url', tx_source_url]
+
+        ingest_enclave = config.get('fog-ingest-enclave-css')
+        if ingest_enclave is not None:
+            wallet_server_command += ['--fog-ingest-enclave-css', ingest_enclave]
+
         if bg:
             wallet_server_command += [
-                '>', str(LOG_LOCATION), '2>&1'
+                '>', config['logfile'], '2>&1'
             ]
-
-        if NETWORK == 'testnet':
-            print('Starting TestNet wallet server...')
-        elif NETWORK == 'mainnet':
-            print('Starting MobileCoin wallet server...')
 
         if self.verbose:
             print(' '.join(wallet_server_command))
 
-        MC_DATA.mkdir(parents=True, exist_ok=True)
-        (MC_DATA / 'ledger-db').mkdir(exist_ok=True)
-        (MC_DATA / 'wallet-db').mkdir(exist_ok=True)
+        print('Starting {}...'.format(Path(config['executable']).name))
 
-        os.environ['RUST_LOG'] = 'info'
-        os.environ['mc_ledger_sync'] = 'info'
+        Path(config['ledger-db']).mkdir(parents=True, exist_ok=True)
+        Path(config['wallet-db']).parent.mkdir(parents=True, exist_ok=True)
+
         if bg:
             subprocess.Popen(' '.join(wallet_server_command), shell=True)
-            print('Started, view log at {}.'.format(LOG_LOCATION))
-            print('Stop server with "mobilecoin stop".')
+            print('Started, view log at {}.'.format(config['logfile']))
+            print('Stop server with "mobcli stop".')
         else:
             subprocess.run(' '.join(wallet_server_command), shell=True)
 
     def stop(self):
         if self.verbose:
             print('Stopping MobileCoin wallet server...')
-        if NETWORK == 'testnet':
-            subprocess.Popen(['killall', '-v', 'full-service-testnet'])
-        elif NETWORK == 'mainnet':
-            subprocess.Popen(['killall', '-v', 'full-service'])
+        subprocess.Popen(['killall', '-v', config['executable']])
 
     def create(self, **args):
         account = self.client.create_account(**args)
-        account_id = account['account_id']
         print('Created a new account.')
-        print(account_id[:6], account['name'])
+        print()
+        _print_account(account)
+        print()
+
+    def rename(self, account_id, name):
+        account = self._load_account_prefix(account_id)
+        old_name = account['name']
+        account_id = account['account_id']
+        account = self.client.update_account_name(account_id, name)
+        print('Renamed account from "{}" to "{}".'.format(
+            old_name,
+            account['name'],
+        ))
+        print()
+        _print_account(account)
+        print()
 
     def import_(self, seed, **args):
-        entropy, block = _load_import(seed)
+        entropy, block, fog_keys = _load_import(seed)
         if args['block'] is None and block is not None:
             args['block'] = block
-        account = self.client.import_account(entropy, **args)
+        account = self.client.import_account(entropy, fog_keys=fog_keys, **args)
         account_id = account['account_id']
         balance = self.client.get_balance_for_account(account_id)
 
@@ -188,38 +192,53 @@ class CommandLineInterface:
         print()
         _print_account(account, balance)
         print()
+        print('Keep the exported seed phrase file safe and private!')
         print('Anyone who has access to the seed phrase can spend all the')
-        print('funds in the account. Keep the exported file safe and private!')
+        print('funds in the account.')
         if not confirm('Really write account seed phrase to a file? (Y/N) '):
             print('Cancelled.')
             return
 
         secrets = self.client.export_account_secrets(account_id)
         filename = 'mobilecoin_seed_phrase_{}.json'.format(account_id[:16])
-        _save_export(account, secrets, filename)
-        print(f'Wrote {filename}.')
+        try:
+            _save_export(account, secrets, filename)
+        except OSError as e:
+            print('Could not write file: {}'.format(e))
+            exit(1)
+        else:
+            print(f'Wrote {filename}.')
 
-    def delete(self, account_id):
+    def qr(self, account_id):
         account = self._load_account_prefix(account_id)
         account_id = account['account_id']
         balance = self.client.get_balance_for_account(account_id)
 
-        amount = pmob2mob(balance['unspent_pmob'])
-        if balance['is_synced'] is True and amount == 0:
-            print('Account {} has 0 MOB.'.format(account_id[:6]))
-        else:
-            print('You are about to delete this account:')
-            print()
-            _print_account(account, balance)
-            print()
-            print('You will lose access to the funds in this account unless you')
-            print('restore it from the seed phrase.')
-            if not confirm('Continue? (Y/N) '):
-                print('Cancelled.')
-                return
+        mob_url = 'mob:///b58/{}'.format(account['main_address'])
+        qr = segno.make(mob_url)
+        qr.terminal()
 
-        self.client.delete_account(account_id)
-        print('Deleted.')
+        print()
+        _print_account(account, balance)
+        print()
+
+    def remove(self, account_id):
+        account = self._load_account_prefix(account_id)
+        account_id = account['account_id']
+        balance = self.client.get_balance_for_account(account_id)
+
+        print('You are about to remove this account:')
+        print()
+        _print_account(account, balance)
+        print()
+        print('You will lose access to the funds in this account unless you')
+        print('restore it from the seed phrase.')
+        if not confirm('Continue? (Y/N) '):
+            print('Cancelled.')
+            return
+
+        self.client.remove_account(account_id)
+        print('Removed.')
 
     def list(self, **args):
         accounts = self.client.get_all_accounts(**args)
@@ -243,58 +262,135 @@ class CommandLineInterface:
 
         print()
 
-    def history(self, account_id):
-        pass
+    def send(self, account_id, amount, to_address, build_only=False):
+        account = self._load_account_prefix(account_id)
+        account_id = account['account_id']
+        balance = self.client.get_balance_for_account(account_id)
+        unspent = pmob2mob(balance['unspent_pmob'])
 
-    def send(self, from_account_id, amount, to_address):
-        account = self._load_account_prefix(from_account_id)
-        from_account_id = account['account_id']
-        amount = Decimal(amount)
+        if amount == "all":
+            amount = unspent - TRANSACTION_FEE
+            total_amount = unspent
+        else:
+            amount = Decimal(amount)
+            total_amount = amount + TRANSACTION_FEE
 
-        print('\n'.join(
-            'Sending {:.4f} MOB from account {} {}',
-            'to address {}.'
-        ).format(
-            amount,
-            from_account_id[:6],
+        if build_only:
+            verb = 'Building transaction for'
+        else:
+            verb = 'Sending'
+
+        print('\n'.join([
+            '{} {} from account {} {}',
+            'to address {}.',
+            'Fee is {}, for a total amount of {}.',
+        ]).format(
+            verb,
+            _format_mob(amount),
+            account_id[:6],
             account['name'],
             to_address,
+            _format_mob(TRANSACTION_FEE),
+            _format_mob(total_amount),
         ))
+
+        if total_amount > unspent:
+            print('\n'.join([
+                'Cannot send this transaction, because the account only',
+                'contains {:.4f} MOB. Try sending all funds by entering amount as "all".',
+            ]).format(unspent))
+            return
+
+        if build_only:
+            tx_proposal = self.client.build_transaction(account_id, amount, to_address)
+            path = Path('tx_proposal.json')
+            if path.exists():
+                print(f'{path} already exists.')
+            else:
+                with path.open('w') as f:
+                    json.dump(tx_proposal, f)
+                print(f'Wrote {path}')
+            return
+
         if not confirm('Confirm? (Y/N) '):
             print('Cancelled.')
             return
 
-        self.client.send_transaction(from_account_id, amount, to_address)
-        print('Sent.')
+        transaction_log = self.client.build_and_submit_transaction(account_id, amount, to_address)
 
-    def prepare():
-        pass
+        print('Sent {:.4f} MOB, with a transaction fee of {:.4f} MOB'.format(
+            pmob2mob(transaction_log['value_pmob']),
+            pmob2mob(transaction_log['fee_pmob']),
+        ))
 
-    def submit():
-        pass
+    def history(self, account_id):
+        account = self._load_account_prefix(account_id)
+        account_id = account['account_id']
 
+        transactions = self.client.get_all_transaction_logs_for_account(account_id)
+        transactions = sorted(
+            transactions.values(),
+            key=lambda t: int(t['finalized_block_index'])
+        )
+        for t in transactions:
+            print()
+            amount = _format_mob(pmob2mob(t['value_pmob']))
+            if t['direction'] == 'tx_direction_received':
+                print('Received {}'.format(amount))
+                print('  at {}'.format(t['assigned_address_id']))
+            elif t['direction'] == 'tx_direction_sent':
+                fee = _format_mob(pmob2mob(t['fee_pmob']))
+                print('Sent {} (fee {})'.format(amount, fee))
+                print('  to {}'.format(t['recipient_address_id']))
+            print('  in block', t['finalized_block_index'])
+        print()
 
 def confirm(message):
     confirmation = input(message)
     return confirmation.lower() in ['y', 'yes']
 
 
-def _print_account(account, balance):
-    account_id = account['account_id']
+def _format_mob(mob):
+    return '{:.4f} MOB'.format(mob)
 
-    total_blocks = int(balance['network_block_index'])
-    offline = (total_blocks == 0)
-    if offline:
-        total_blocks = balance['local_block_index']
+
+def _print_account(account, balance=None):
+    account_id = account['account_id']
 
     print(account_id[:6], account['name'])
     print('  address', account['main_address'])
-    print('  {:.4f} MOB ({}/{} blocks synced) {}'.format(
-        pmob2mob(balance['unspent_pmob']),
-        balance['account_block_index'],
-        total_blocks,
-        ' [offline]' if offline else '',
-    ))
+
+    if balance is not None:
+        total_blocks = int(balance['network_block_index'])
+        offline = (total_blocks == 0)
+        if offline:
+            total_blocks = balance['local_block_index']
+        print('  {} ({}/{} blocks synced) {}'.format(
+            _format_mob(pmob2mob(balance['unspent_pmob'])),
+            balance['account_block_index'],
+            total_blocks,
+            ' [offline]' if offline else '',
+        ))
+
+
+def _print_txo(txo, received=False):
+    print(txo)
+    to_address = txo['assigned_address']
+    if received:
+        verb = 'Received'
+    else:
+        verb = 'Spent'
+    print('  {} {}'.format(verb, _format_mob(pmob2mob(txo['value_pmob']))))
+    if received:
+        if int(txo['subaddress_index']) == 1:
+            print('    as change')
+        else:
+            print('    at subaddress {}, {}'.format(
+                txo['subaddress_index'],
+                to_address,
+            ))
+    else:
+        print('    to unknown address')
 
 
 def _load_import(seed):
@@ -303,21 +399,35 @@ def _load_import(seed):
         b = bytes.fromhex(seed)
         if len(b) == 32:
             entropy = b.hex()
-            return entropy, None
+            return entropy, None, None
     except ValueError:
         pass
 
     # Try to interpret it as a BIP39 mnemonic.
     try:
         entropy = Mnemonic('english').to_entropy(seed).hex()
-        return entropy, None
+        return entropy, None, None
     except (ValueError, LookupError):
         pass
 
-    # Last chance, try to open it as a JSON filename.
+    # Try to open it as a JSON filename.
     with open(seed) as f:
         data = json.load(f)
-    return data['root_entropy'], data['first_block_index']
+        fog_keys = {}
+        for field in [
+            'fog_report_url',
+            'fog_report_id',
+            'fog_authority_spki',
+        ]:
+            value = data['account_key'].get(field)
+            if value is not None:
+                fog_keys[field] = value
+
+    return (
+        data['root_entropy'],
+        data.get('first_block_index'),
+        fog_keys,
+    )
 
 
 def _save_export(account, secrets, filename):
@@ -332,6 +442,11 @@ def _save_export(account, secrets, filename):
         "account_key": secrets['account_key'],
         "first_block_index": account['first_block_index'],
     }
-    with open(filename, 'w') as f:
+
+    path = Path(filename)
+    if path.exists():
+        raise OSError('File exists.')
+
+    with path.open('w') as f:
         json.dump(export_data, f, indent=4)
         f.write('\n')
